fix(spark): use DoubleType for price/quantity instead of FloatType
FloatType (32-bit) can't exactly represent most 2-decimal values, so
F.round(price, 2) was silently losing precision again as soon as the
rounded value got stored back into a 32-bit float column (e.g. 187.42
became 187.4199981689453). BigQuery only has a single float type
(FLOAT64/double) anyway, so FloatType never bought anything on the
write side. Switching both fields to DoubleType in silver and gold
(kept in sync, as usual) fixes the rounding and matches BigQuery's
native type.

+++ python/spark_gold.py
@@ -1,5 +1,5 @@
 from pyspark.sql import SparkSession
-from pyspark.sql.types import StringType, TimestampType, StructField, StructType, FloatType, IntegerType
+from pyspark.sql.types import StringType, TimestampType, StructField, StructType, DoubleType, IntegerType
 
 try:
     spark = (
@@ -14,9 +14,9 @@ try:
         [
             StructField("id", IntegerType()),
             StructField("symbol", StringType()),
-            StructField("price", FloatType()),
+            StructField("price", DoubleType()),
             StructField("broker", StringType()),
-            StructField("quantity", FloatType()),
+            StructField("quantity", DoubleType()),
             StructField("currency", StringType()),
             StructField("exchange", StringType()),
             StructField("timestamp", TimestampType())
